Move v2-specific updater code to its own file, like for v3
